requirements.txt

boto3>=1.26.0
requests>=2.28.0
PyJWT>=2.6.0
pandas>=1.5.0
python-dotenv>=1.0.0
pyarrow>=10.0.0
cryptography>=38.0.0